backend/requirements.txt
fastapi==0.68.1
asyncpg==0.24.0
sqlmodel==0.0.4
uvicorn==0.15.0
alembic==1.7.1
passlib==1.7.4
python-jose==3.3.0
bcrypt==3.2.0
python-multipart==0.0.5
asyncio==3.4.3